+++ tests/test_C1_verification.py
@@ -312,19 +312,19 @@ def test_check_feedin_tariff_vs_levelized_cost_of_generation_of_production_non_d
     ), f"If the capacity of a production asset is not optimized and the feed-in tariff is greater than the expected lcoe a debug msg should be logged."
 
 
-def test_check_maximum_emissions_constraint_no_warning_no_constraint(caplog):
+def test_check_feasibility_of_maximum_emissions_constraint_no_warning_no_constraint(caplog):
     dict_values = {
         CONSTRAINTS: {MAXIMUM_EMISSIONS: {VALUE: None}},
     }
     # logging.warning
     with caplog.at_level(logging.WARNING):
-        C1.check_maximum_emissions_constraint(dict_values)
+        C1.check_feasibility_of_maximum_emissions_constraint(dict_values)
     assert (
         caplog.text == ""
     ), f"If the maximum emissions constraint is not used, no msg should be logged."
 
 
-def test_check_maximum_emissions_constraint_no_warning_although_emission_contraint(
+def test_check_feasibility_of_maximum_emissions_constraint_no_warning_although_emission_constraint(
     caplog,
 ):
     dict_values = {
@@ -339,13 +339,13 @@ def test_check_maximum_emissions_constraint_no_warning_although_emission_contrai
     }
     # logging.warning
     with caplog.at_level(logging.WARNING):
-        C1.check_maximum_emissions_constraint(dict_values)
+        C1.check_feasibility_of_maximum_emissions_constraint(dict_values)
     assert (
         caplog.text == ""
     ), f"If the maximum emissions constraint is used and one production asset with zero emissions has a capacity to be optimized and a maximum capacity of None, no msg should be logged."
 
 
-def test_check_maximum_emissions_constraint_maximumcap(caplog):
+def test_check_feasibility_of_maximum_emissions_constraint_maximumcap(caplog):
     dict_values = {
         CONSTRAINTS: {MAXIMUM_EMISSIONS: {VALUE: 1000}},
         ENERGY_PRODUCTION: {
@@ -358,14 +358,14 @@ def test_check_maximum_emissions_constraint_maximumcap(caplog):
     }
     # logging.warning
     with caplog.at_level(logging.WARNING):
-        C1.check_maximum_emissions_constraint(dict_values)
+        C1.check_feasibility_of_maximum_emissions_constraint(dict_values)
     assert (
         "When the maximum emissions constraint is used and no production asset with zero emissions"
         in caplog.text
     ), f"If the maximum emissions constraint is used and no production asset with zero emissions, capacity to be optimized and maximum capacity of None exists, a warning msg should be logged."
 
 
-def test_check_maximum_emissions_constraint_optimizeCap_is_False(caplog):
+def test_check_feasibility_of_maximum_emissions_constraint_optimizeCap_is_False(caplog):
     dict_values = {
         CONSTRAINTS: {MAXIMUM_EMISSIONS: {VALUE: 1000}},
         ENERGY_PRODUCTION: {
@@ -378,14 +378,14 @@ def test_check_maximum_emissions_constraint_optimizeCap_is_False(caplog):
     }
     # logging.warning
     with caplog.at_level(logging.WARNING):
-        C1.check_maximum_emissions_constraint(dict_values)
+        C1.check_feasibility_of_maximum_emissions_constraint(dict_values)
     assert (
         "When the maximum emissions constraint is used and no production asset with zero emissions"
         in caplog.text
     ), f"If the maximum emissions constraint is used and no production asset with zero emissions, capacity to be optimized and maximum capacity of None exists, a warning msg should be logged."
 
 
-def test_check_maximum_emissions_constraint_no_zero_emission_asset(caplog):
+def test_check_feasibility_of_maximum_emissions_constraint_no_zero_emission_asset(caplog):
     dict_values = {
         CONSTRAINTS: {MAXIMUM_EMISSIONS: {VALUE: 1000}},
         ENERGY_PRODUCTION: {
@@ -398,7 +398,7 @@ def test_check_maximum_emissions_constraint_no_zero_emission_asset(caplog):
     }
     # logging.warning
     with caplog.at_level(logging.WARNING):
-        C1.check_maximum_emissions_constraint(dict_values)
+        C1.check_feasibility_of_maximum_emissions_constraint(dict_values)
     assert (
         "When the maximum emissions constraint is used and no production asset with zero emissions"
         in caplog.text